Minimapa no desaparace bajo textura

+++ inc/cub_structs.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   cub_structs.h                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: erivero- <erivero-@student.42.fr>          +#+  +:+       +#+        */
+/*   By: marirodr <marirodr@student.42malaga.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/23 10:52:50 by marirodr          #+#    #+#             */
-/*   Updated: 2024/01/04 13:19:01 by erivero-         ###   ########.fr       */
+/*   Updated: 2024/01/08 18:56:41 by marirodr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -64,7 +64,7 @@ typedef struct s_player
 	int			color;
 	double		mov_speed;
 	double		rot_speed;
-	float		angle; //(?)
+	int			angle; //(?)
 	int			size;
 }	t_player;
